Slight changes

+++ time/src/main/java/com/stardevllc/starlib/time/TimeFormat.java
@@ -53,7 +53,7 @@ public class TimeFormat {
                 continue;
             }
 
-            long unitLength = unit.fromMillis(totalTime);
+            long unitLength = (long) unit.fromMillis(totalTime);
             //totalTime -= unit.toMillis(unitLength);
             totalTime = totalTime % unit.getMsPerUnit();
